+++ portal/app_settings.py
@@ -49,9 +49,6 @@ RECAPTCHA_PUBLIC_KEY = getattr(settings, "RECAPTCHA_PUBLIC_KEY", None)
 
 DEBUG = getattr(settings, "DEBUG", False)
 
-# The name of the google app engine service the application is running on, local otherwise
-MODULE_NAME = getattr(settings, "MODULE_NAME", "local")
-
 # The permission function for checking if the request is coming from a cron job
 IS_CLOUD_SCHEDULER_FUNCTION = getattr(
     settings, "IS_CLOUD_SCHEDULER_FUNCTION", lambda _: False