+++ portal/urls.py
@@ -60,7 +60,6 @@ from wagtail.documents import urls as wagtaildocs_urls
 
 from portal.views.about import about
 from portal.views.admin import (
-    AdminLoginView,
     AdminChangePasswordView,
     AdminChangePasswordDoneView,
     aggregated_data,
@@ -198,11 +197,6 @@ urlpatterns = [
         RedirectView.as_view(url="/static/portal/img/favicon.ico", permanent=True),
     ),
     url(
-        r"^administration/login/$",
-        AdminLoginView.as_view(),
-        name="administration_login",
-    ),
-    url(
         r"^administration/password_change/$",
         AdminChangePasswordView.as_view(),
         name="administration_password_change",
@@ -330,7 +324,7 @@ urlpatterns = [
     url(r"^about", about, name="about"),
     url(r"^help/$", contact, name="help"),
     url(r"^terms", terms, name="terms"),
-    url(r"^privacy-policy", privacy_policy, name="privacy_policy"),
+    url(r"^privacy-policy/$", privacy_policy, name="privacy_policy"),
     url(r"^teach/materials/$", materials, name="materials"),
     url(r"^teach/kurono_teaching_packs$", kurono_teaching_packs, name="kurono_packs"),
     url(